исправлены важные ошибки:
* отрисовка кластеров на графике
* жестко задано количество разбиений, как степень 2. тогда не может
возникнуть ситуации, что число кластеров уменьшится при увеличении
количества разбиений.

+++ src/distanceelement.h
@@ -75,7 +75,7 @@ public:
     {
         QVector<int> res;
         for (int i = 0; i < coeffs.size(); i++) {
-            res << coeffs.at(i)/((fabs(max)+fabs(min))/(double)partsCnt);
+            res << coeffs.at(i)/((fabs(max)+fabs(min))/partsCnt);
         }
         return res;
     }
